CardChart component made
